Change behaviour during raw prompt extraction (#989)
This PR suggest a fix on a bug that when `_switch_chat_template()`
method is called.

According to

https://github.com/volcengine/verl/blob/main/verl/utils/dataset/rl_dataset.py#L222
`data.non_tensor_batch['raw_prompt'][i]` is already a list if
`data.return_raw_chat=True`.

Calling `.tolist()` again will result an error. Now we check if it is a
list before run this method.